backend/requirements.txt
django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-storages==1.14.2
boto3==1.34.11
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
reportlab==4.0.9
django-filter==23.5
drf-spectacular==0.27.1
gunicorn==21.2.0
whitenoise==6.6.0
resend==2.2.0
